Add insights service provider
